\docType{package}
\name{WMTools-package}
\alias{WMTools}
\alias{WMTools-package}
\title{Tools for simulating activations in Warmachine(TM)}
\description{
Simulate ranged and melee attacks in the game of
Warmachine(TM)
}
\details{
\tabular{ll}{ Package: \tab WMTools \cr Type: \tab Package
\cr Version: \tab 0.1 \cr Date: \tab 2014-09-26 \cr
Lazyload: \tab yes \cr }
}
\keyword{game}
\keyword{package}
\keyword{simulation}

